% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{add_no_selected_choices}
\alias{add_no_selected_choices}
\title{Add empty choice to choices selected}
\usage{
add_no_selected_choices(x, multiple = FALSE)
}
\arguments{
\item{x}{(\code{choices_selected}) object.}

\item{multiple}{(\code{logical(1)}) whether multiple selections are allowed or not.}
}
\value{
\code{choices_selected} object with an empty option added to the choices.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
