% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_datasets.R
\name{are_needed_keys_provided}
\alias{are_needed_keys_provided}
\title{Checks whether the provided slices have the corresponding join keys}
\usage{
are_needed_keys_provided(join_keys, merged_selector_list)
}
\arguments{
\item{join_keys}{(\code{join_keys}) the provided join keys.}

\item{merged_selector_list}{(\code{list}) the specification of datasets' slices to merge.}
}
\value{
\code{TRUE} if all pairs of the slices have the corresponding keys and
\code{FALSE} otherwise.
}
\description{
Checks whether the provided slices have the corresponding join keys
}
\note{
\code{merged_selector_list} contains a list of descriptions of data frame slices;
each coming from a single dataset. This function checks whether all pairs
of the datasets have the join keys needed to merge the slices.
}
\keyword{internal}
