% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-badge_dropdown.R
\name{badge_dropdown}
\alias{badge_dropdown}
\title{Drop-down badge}
\usage{
badge_dropdown(id, label, content)
}
\arguments{
\item{id}{(\code{character(1)}) shiny module's id}

\item{label}{(\code{shiny.tag}) Label displayed on a badge.}

\item{content}{(\code{shiny.tag}) Content of a drop-down.}
}
\description{
Drop-down button in a form of a badge with \code{bg-primary} as default style
Clicking badge shows a drop-down containing any \code{HTML} element. Folded drop-down
doesn't trigger display output which means that items rendered using \verb{render*}
will be recomputed only when drop-down is show.
}
\keyword{internal}
