% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_condition_range}
\alias{call_condition_range}
\title{\code{numeric} range condition call}
\usage{
call_condition_range(varname, range)
}
\arguments{
\item{varname}{(\code{name} or \code{character(1)})

name of the variable}

\item{range}{(\code{numeric(2)})

range of the variable}
}
\value{
a \code{call}
}
\description{
Compose \code{numeric} range condition call from inputs
}
\examples{
# use non-exported function from teal.transform
call_condition_range <- getFromNamespace("call_condition_range", "teal.transform")
call_extract_list <- getFromNamespace("call_extract_list", "teal.transform")

call_condition_range("AGE", range = c(1, 2))
call_condition_range(as.name("AGE"), range = c(-1.2, 2.1))
call_condition_range(
  call_extract_list("ADSL", "AGE"),
  range = c(-1.2, 2.1)
)
}
\keyword{internal}
