% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_extract_array}
\alias{call_extract_array}
\title{Get call to subset and select array}
\usage{
call_extract_array(dataname = ".", row = NULL, column = NULL, aisle = NULL)
}
\arguments{
\item{dataname}{(\code{character(1)} or \code{name}).}

\item{row}{(optional \code{name} or \code{call} or \code{logical} or \code{integer} or \code{character})
name of the \code{row} or condition.}

\item{column}{(optional \code{name}, \code{call}, \code{logical}, \code{integer} or \code{character})
name of the \code{column} or condition.}

\item{aisle}{(optional \code{name}, \code{call}, \code{logical}, \code{integer} or \code{character})
name of the \code{row} or condition.}
}
\value{
\code{\link[=Extract]{Extract()}} \code{call} for 3-dimensional array in \code{x[i, j, k]} notation.
}
\description{
Get call to subset and select array
}
\examples{
# use non-exported function from teal.transform
call_extract_array <- getFromNamespace("call_extract_array", "teal.transform")
call_condition_choice <- getFromNamespace("call_condition_choice", "teal.transform")

call_extract_array(
  dataname = "my_array",
  row = call_condition_choice("my_array$SEX", "M"),
  column = call("c", "SEX", "AGE"),
  aisle = "RNAseq_rnaaccess"
)
call_extract_array(
  "mae_object",
  column = call_condition_choice("SEX", "M")
)
}
\keyword{internal}
