% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_extract_list}
\alias{call_extract_list}
\title{Compose extract call with \code{$} operator}
\usage{
call_extract_list(dataname, varname, dollar = TRUE)
}
\arguments{
\item{dataname}{(\code{character(1)} or \code{name})
name of the object}

\item{varname}{(\code{character(1)} or \code{name})
name of the slot in data}

\item{dollar}{(\code{logical(1)})\cr
whether returned call should use \code{$} or \code{[[} operator}
}
\value{
\code{$} or \code{[[} call
}
\description{
Compose extract call with \code{$} operator
}
\examples{
# use non-exported function from teal.transform
call_extract_list <- getFromNamespace("call_extract_list", "teal.transform")

call_extract_list("ADSL", "SEX")
call_extract_list("ADSL", "named element")
call_extract_list(as.name("ADSL"), as.name("AGE"))
call_extract_list(as.name("weird name"), as.name("AGE"))
call_extract_list(as.name("ADSL"), "AGE", dollar = FALSE)
}
\keyword{internal}
