% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_uses_magrittr_pipe}
\alias{call_uses_magrittr_pipe}
\title{Check if a call or list of calls uses the pipe operator (\%>\%)}
\usage{
call_uses_magrittr_pipe(x)
}
\arguments{
\item{x}{(\code{call}, \code{name}, or \code{list} of calls) The call(s) to check.}
}
\value{
\code{logical(1)} \code{TRUE} if \verb{\%>\%} is found, \code{FALSE} otherwise.
}
\description{
Recursively searches through a call or list of calls to determine
if the pipe operator \verb{\%>\%} is used anywhere.
}
\keyword{internal}
