% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_with_colon}
\alias{call_with_colon}
\title{Create a call using a function in a given namespace}
\usage{
call_with_colon(name, ..., unlist_args = list())
}
\arguments{
\item{name}{\code{character} function name, possibly using namespace colon \code{::}, also
works with \code{:::} (sometimes needed, but strongly discouraged).}

\item{...}{arguments to pass to function with name \code{name}.}

\item{unlist_args}{\code{list} extra arguments passed in a single list,
avoids the use of \code{do.call} with this function.}
}
\value{
\code{call}.
}
\description{
The dot arguments in \code{...} need to be quoted because they will be evaluated otherwise.
}
\keyword{internal}
