% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_module.R
\name{check_data_extract_spec}
\alias{check_data_extract_spec}
\title{Check data extract specification}
\usage{
check_data_extract_spec(data_extract_spec)
}
\arguments{
\item{data_extract_spec}{(\code{list}) of \code{data_extract_spec}.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{data_extract_spec}
parameter, invisibly and unchanged.
}
\description{
Check data extract specification
}
\keyword{internal}
