% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_module.R
\name{check_data_extract_spec_react}
\alias{check_data_extract_spec_react}
\title{Function to check data_extract_specs}
\usage{
check_data_extract_spec_react(datasets, data_extract)
}
\arguments{
\item{datasets}{(\code{FilteredData}) the object created using the \code{teal} API.}

\item{data_extract}{(\code{list}) the output of the \code{data_extract} module.}
}
\value{
\code{NULL}.
}
\description{
Checks if \code{dataname} argument exists as a dataset.
Checks if selected or filter columns exist within the datasets. Throws a \code{shiny}
validation error if the above requirements are not met.
}
\keyword{internal}
