% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_datasets.R
\name{check_data_merge_selectors}
\alias{check_data_merge_selectors}
\title{Validate data_extracts in merge_datasets}
\usage{
check_data_merge_selectors(selector_list)
}
\arguments{
\item{selector_list}{(\code{reactive})
output from \code{\link[=data_extract_multiple_srv]{data_extract_multiple_srv()}} or a reactive named list of
outputs from \code{\link[=data_extract_srv]{data_extract_srv()}}.
When using a reactive named list, the names must be identical to the shiny
ids of the respective
\code{\link[=data_extract_ui]{data_extract_ui()}}.}
}
\value{
\code{NULL} if check is successful and \code{shiny} validate error otherwise.
}
\description{
Validate selected inputs from data_extract before passing to data_merge to avoid
\code{dplyr} errors or unexpected results.
}
\keyword{internal}
