% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data_utils.R
\name{check_merge_function}
\alias{check_merge_function}
\title{Check if the merge function is valid}
\usage{
check_merge_function(merge_function)
}
\arguments{
\item{merge_function}{(\code{character}) merge function name.}
}
\value{
Raises an error when check fails, otherwise, it returns \code{NULL}, invisibly.
}
\description{
Check if the merge function is valid
}
\keyword{internal}
