% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_no_multiple_selection}
\alias{check_no_multiple_selection}
\title{Checks that the extract_input specification does not allow multiple
selection}
\usage{
check_no_multiple_selection(extract_input)
}
\arguments{
\item{extract_input}{\code{data_extract_spec} a list of \code{data_extract_spec} or NULL

Stops if condition not met}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
