% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector}
\alias{check_selector}
\title{Check selector}
\usage{
check_selector(selector)
}
\arguments{
\item{selector}{(\code{list}) of selector elements generated by \code{data_extract_srv}.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{selector}
parameter, invisibly and unchanged.
}
\description{
Check selector
}
\keyword{internal}
