% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector_dataname}
\alias{check_selector_dataname}
\title{Check selector \code{dataname} element}
\usage{
check_selector_dataname(dataname)
}
\arguments{
\item{dataname}{(\code{character(1)}) selector element.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{dataname}
parameter, invisibly and unchanged.
}
\description{
Check selector \code{dataname} element
}
\keyword{internal}
