% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector_filters}
\alias{check_selector_filters}
\title{Check selector filters element}
\usage{
check_selector_filters(filters)
}
\arguments{
\item{filters}{(\code{list}) selector element generated by \code{data_extract_srv}.}
}
\value{
Raises an error when the check fails, otherwise it returns \code{NULL}, invisibly.
}
\description{
Check selector filters element
}
\keyword{internal}
