% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector_internal_id}
\alias{check_selector_internal_id}
\title{Check selector internal_id element}
\usage{
check_selector_internal_id(internal_id)
}
\arguments{
\item{internal_id}{(\code{character(1)}) selector element generated by \code{data_extract_srv}.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{internal_id}
parameter, invisibly and unchanged.
}
\description{
Check selector internal_id element
}
\keyword{internal}
