% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector_keys}
\alias{check_selector_keys}
\title{Check selector keys element}
\usage{
check_selector_keys(keys)
}
\arguments{
\item{keys}{(\code{character}) selector element generated by \code{data_extract_srv}.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{keys}
parameter, invisibly and unchanged.
}
\description{
Check selector keys element
}
\keyword{internal}
