% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector_reshape}
\alias{check_selector_reshape}
\title{Check selector reshape element}
\usage{
check_selector_reshape(reshape)
}
\arguments{
\item{reshape}{(\code{logical(1)}) selector element generated by \code{data_extract_srv}.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{reshape}
parameter, invisibly and unchanged.
}
\description{
Check selector reshape element
}
\keyword{internal}
