% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selector.R
\name{check_selector_select}
\alias{check_selector_select}
\title{Check selector select element}
\usage{
check_selector_select(select)
}
\arguments{
\item{select}{(\code{character}) selector element generated by \code{data_extract_srv}.}
}
\value{
Raises an error when check fails, otherwise, it returns the \code{select}
parameter, invisibly and unchanged.
}
\description{
Check selector select element
}
\keyword{internal}
