% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_module.R
\name{cond_data_extract_single_ui}
\alias{cond_data_extract_single_ui}
\title{Creates a panel that displays (with filter and column selection)
conditionally on \code{input[ns("dataset")] == dataname}}
\usage{
cond_data_extract_single_ui(ns, single_data_extract_spec)
}
\arguments{
\item{ns}{(\code{function}) the shiny namespace function.}

\item{single_data_extract_spec}{(\code{data_extract_spec}) the specification
for extraction of data during the application initialization.

Generated by \code{\link[=data_extract_spec]{data_extract_spec()}}.}
}
\value{
\code{shiny.tag} with the HTML code for the panel.
}
\description{
Creates a panel that displays (with filter and column selection)
conditionally on \code{input[ns("dataset")] == dataname}
}
\keyword{internal}
