% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_teal_data}
\alias{convert_teal_data}
\title{Ensures datasets is a list of reactive expression}
\usage{
convert_teal_data(datasets)
}
\arguments{
\item{datasets}{(\code{reactive} or \code{teal_data} or \code{list}) of \code{data.frame}
wrapped or not in a reactive expression.}
}
\value{
List of \code{reactive} expressions that contains all the individual \code{datasets}.
}
\description{
Ensures datasets is a list of reactive expression
}
\keyword{internal}
