% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_filter_module.R
\name{data_extract_filter_ui}
\alias{data_extract_filter_ui}
\title{Returns a \code{shiny.tag} object with the UI for a \code{filter_spec} object}
\usage{
data_extract_filter_ui(filter, id = "filter")
}
\arguments{
\item{filter}{(\code{filter_spec}) the object generated with \code{\link[=filter_spec]{filter_spec()}}.}

\item{id}{(\code{character(1)}) the shiny \code{inputId} for the generated \code{shiny.tag}.}
}
\value{
\code{shiny.tag} defining the \code{filter_spec}'s UI element.
}
\description{
Returns a \code{shiny.tag} object with the UI for a \code{filter_spec} object
}
\details{
Creates two \code{optionSelectInput} elements (one for column and one for values) based
on a definition of a \code{\link[=filter_spec]{filter_spec()}} object.
}
\keyword{internal}
