% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_single_module.R
\name{data_extract_single_srv}
\alias{data_extract_single_srv}
\title{The server function for a single \code{data_extract_spec} object}
\usage{
data_extract_single_srv(id, datasets, single_data_extract_spec)
}
\arguments{
\item{id}{(\code{character}) id string.}

\item{datasets}{(\verb{named list}) a list of reactive \code{data.frame} type objects.}

\item{single_data_extract_spec}{(\code{data_extract_spec}) the
\code{\link[=data_extract_spec]{data_extract_spec()}} object to handle.}
}
\value{
\code{NULL}.
}
\description{
The server function for a single \code{data_extract_spec} object
}
\details{
The Shiny server function for handling a single \link{data_extract_spec} object.
}
\keyword{internal}
