% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge_module.R
\name{data_merge_module}
\alias{data_merge_module}
\title{Data merge module}
\usage{
data_merge_module(
  datasets,
  data_extract,
  merge_function = "dplyr::full_join",
  anl_name = "ANL",
  id = "merge_id"
)
}
\arguments{
\item{datasets}{(\code{FilteredData})\cr
object containing data, see \link[teal.slice:FilteredData]{teal.slice::FilteredData} for more.}

\item{data_extract}{(named \code{list} of \code{data_extract_spec})\cr}

\item{merge_function}{(\code{character(1)})\cr
A character string of a function that
accepts the arguments \code{x}, \code{y} and \code{by} to perform the merging of datasets.}

\item{anl_name}{(\code{character(1)})\cr
Name of the analysis dataset.}

\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
reactive expression with output from \code{\link[=data_merge_srv]{data_merge_srv()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function was a convenient wrapper to combine \code{data_extract_multiple_srv()} and
\code{data_merge_srv()} when no additional processing is required.
}
\seealso{
\code{\link[=data_merge_srv]{data_merge_srv()}}
}
