% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge_module.R
\name{data_merge_srv}
\alias{data_merge_srv}
\title{Data merge module server}
\usage{
data_merge_srv(
  id = "merge_id",
  selector_list,
  datasets,
  merge_function = "dplyr::full_join",
  anl_name = "ANL"
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{selector_list}{(\code{reactive})\cr
output from \code{\link[=data_extract_multiple_srv]{data_extract_multiple_srv()}} or a reactive named list of outputs from \code{\link[=data_extract_srv]{data_extract_srv()}}.
When using a reactive named list, the names must be identical to the shiny ids of the
respective \code{\link[=data_extract_ui]{data_extract_ui()}}.}

\item{datasets}{(\code{FilteredData})\cr
object containing data (see \code{teal.slice::FilteredData}).}

\item{merge_function}{(\code{character(1)} or \code{reactive})\cr
A character string of a function that accepts the arguments
\code{x}, \code{y} and \code{by} to perform the merging of datasets.}

\item{anl_name}{(\code{character(1)})\cr
Name of the analysis dataset.}
}
\value{
reactive expression with output from \link{merge_datasets}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
When additional processing of the \code{data_extract} list input was required, \code{data_merge_srv()} could be combined
with \code{data_extract_multiple_srv()} or \code{data_extract_srv()} to influence the \code{selector_list} input.
}
\seealso{
\code{\link[=data_extract_srv]{data_extract_srv()}}
}
