% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_datanames.R
\name{datanames_input}
\alias{datanames_input}
\title{Available datasets input}
\usage{
datanames_input(data_extracts)
}
\arguments{
\item{data_extracts}{list of data extracts for single variable}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates \link[shiny]{helpText} with the names of datasets
available for current module.
}
