% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-picks.R
\name{.is_delayed}
\alias{.is_delayed}
\title{Is picks delayed}
\usage{
.is_delayed(x)
}
\arguments{
\item{x}{(\code{list}, \verb{list of picks}, \code{picks}, \code{pick}, \verb{$choices}, \verb{$selected})}
}
\description{
Determine whether list of picks/picks or pick are delayed.
When \code{\link[dplyr:pick]{dplyr::pick()}} is created it could be either:
\itemize{
\item \code{quosure} when \code{tidyselect} helper used (delayed)
\item \code{function} when predicate function provided (delayed)
\item \code{atomic} when vector of choices/selected provided (eager)
}
}
\keyword{internal}
