% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-types.R
\name{.is_tidyselect}
\alias{.is_tidyselect}
\title{Is an object created using tidyselect}
\usage{
.is_tidyselect(x)
}
\arguments{
\item{x}{\code{choices} or \code{selected}}
}
\value{
\code{logical(1)}
}
\description{
\code{choices} and \code{selected} can be provided using \code{tidyselect}, (e.g. \code{\link[tidyselect:everything]{tidyselect::everything()}}
\code{\link[tidyselect:where]{tidyselect::where()}}, \code{\link[tidyselect:starts_with]{tidyselect::starts_with()}}). These functions can't be called
independently but rather as an argument of function which consumes them.
\code{.is_tidyselect} safely determines if \code{x} can be evaluated with \code{tidyselect::eval_select()}
}
\keyword{internal}
