% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-module_picks.R
\name{.picker_icon}
\alias{.picker_icon}
\title{\code{pickerInput} choices icons}
\usage{
.picker_icon(x)
}
\arguments{
\item{x}{(\code{any}) object which class will determine icon}
}
\value{
html-tag in form of \code{character(1)}
}
\description{
Icons describing a class of the choice
}
\keyword{internal}
