% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-extract.R
\name{extract}
\alias{extract}
\title{Internal method to extract data from different objects}
\usage{
extract(x, variable, ...)
}
\arguments{
\item{x}{Object from which a subset/element is required.}

\item{variable}{Name of the element to be extracted.}

\item{...}{Other arguments passed to the specific method.}
}
\description{
Required to resolve a specification into something usable (by comparing with the existing data).
Required by merging data based on a resolved specification.
}
\examples{
extract(iris, "Sepal.Length")
}
