% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{get_anl_relabel_call}
\alias{get_anl_relabel_call}
\title{Gets the relabel call}
\usage{
get_anl_relabel_call(columns_source, data, anl_name = "ANL")
}
\arguments{
\item{columns_source}{\code{named list}\cr
where names are column names, values are labels + additional attribute dataname}

\item{data}{(named \code{list})\cr of datasets.}

\item{anl_name}{(\code{character(1)})\cr
Name of the analysis dataset.}
}
\value{
(\code{call}) to relabel \code{dataset} and assign to \code{anl_name}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
