% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_dplyr_call_data}
\alias{get_dplyr_call_data}
\title{Aggregates data extract selectors}
\usage{
get_dplyr_call_data(selector_list, join_keys = teal.data::join_keys())
}
\arguments{
\item{selector_list}{(\code{reactive})
output from \code{\link[=data_extract_multiple_srv]{data_extract_multiple_srv()}} or a reactive named list of
outputs from \code{\link[=data_extract_srv]{data_extract_srv()}}.
When using a reactive named list, the names must be identical to the shiny
ids of the respective
\code{\link[=data_extract_ui]{data_extract_ui()}}.}

\item{join_keys}{(\code{join_keys}) nested list of keys used for joining.}
}
\value{
(\code{list}) simplified selectors with aggregated set of filters,
selections, reshapes etc. All necessary data for merging.
}
\description{
Simplifies \code{selector_list} into aggregated list with one element per
same selector - same dataset, same filter configuration and same reshape status.
}
\keyword{internal}
