% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{get_dropped_filters}
\alias{get_dropped_filters}
\title{Names of filtered-out filters dropped from selection}
\usage{
get_dropped_filters(selector)
}
\arguments{
\item{selector}{one element of selector_list obtained by \code{get_dplyr_call_data}.}
}
\value{
Vector of \code{character} names of the filters which should be dropped from select call.
}
\description{
Names of filtered-out filters dropped from selection
}
\details{
Names of filtered-out filters dropped from automatic selection
(key vars are automatically included in select).
Dropped filter is filter which became not unique for all observations.
This means that if variable is filtered to just one level,
it's not a key anymore.

Other variables used in filter should also be dropped from automatic
selection, unless they have been selected.
}
\keyword{internal}
