% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_filter_call}
\alias{get_filter_call}
\title{Returns \code{dplyr} filter call}
\usage{
get_filter_call(filter, dataname = NULL, datasets = NULL)
}
\arguments{
\item{filter}{(\code{list}) Either list of lists or list with \code{select} and \code{selected} items.}

\item{dataname}{(\code{NULL} or \code{character}) name of dataset.}

\item{datasets}{(\code{NULL} or \code{named `list`}).}
}
\value{
(\code{call}) \code{dplyr} filter call
}
\description{
Returns \code{dplyr} filter call
}
\examples{
teal.transform:::get_filter_call(
  filter = list(list(columns = "SEX", selected = list(NA, "F", "M")))
)
teal.transform:::get_filter_call(filter = list(
  list(columns = "SEX", selected = list(NA, "F", "M")),
  list(columns = "VAR", selected = list("LEVEL1", "LEVEL2"))
))
}
\keyword{internal}
