% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_filter_module.R
\name{get_initial_filter_values}
\alias{get_initial_filter_values}
\title{Returns the initial values for the \code{vals} widget
of a \code{filter_spec} object.}
\usage{
get_initial_filter_values(filter, datasets)
}
\arguments{
\item{filter}{(\code{filter_spec}) the filter generated by a call to \code{\link[=filter_spec]{filter_spec()}}}

\item{datasets}{(\code{FilteredData}) the datasets}
}
\value{
\code{list} of two values: \code{col} and \code{vals}
}
\description{
Returns the initial values for the \code{vals} widget
of a \code{filter_spec} object.
}
\examples{
filtered_data <- teal.slice:::FilteredData$new()$set_dataset(
  teal.data::dataset("iris", utils::head(iris))
)
filter <- filter_spec(vars = colnames(iris)[1])
filter$dataname <- "iris"
shiny::isolate(
  teal.transform:::get_initial_filter_values(filter = filter, datasets = filtered_data)
)
}
\keyword{internal}
