% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{get_merge_key_i}
\alias{get_merge_key_i}
\title{Gets keys needed for join call of two selectors}
\usage{
get_merge_key_i(
  selector_list,
  idx,
  dplyr_call_data = get_dplyr_call_data(selector_list)
)
}
\arguments{
\item{selector_list}{(\code{reactive})
output from \code{\link[=data_extract_multiple_srv]{data_extract_multiple_srv()}} or a reactive named list of
outputs from \code{\link[=data_extract_srv]{data_extract_srv()}}.
When using a reactive named list, the names must be identical to the shiny
ids of the respective
\code{\link[=data_extract_ui]{data_extract_ui()}}.}

\item{idx}{(\code{integer}) optional, current selector index in all selectors list.}

\item{dplyr_call_data}{(\code{list}) simplified selectors with aggregated set of filters.}
}
\value{
\code{character} list of keys.
}
\description{
Gets keys needed for join call of two selectors
}
\keyword{internal}
