% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_pivot_longer_col}
\alias{get_pivot_longer_col}
\title{Get pivot longer  columns}
\usage{
get_pivot_longer_col(selector)
}
\arguments{
\item{selector}{one element of selector_list obtained by \code{get_dplyr_call_data}.}
}
\value{
A \code{character} vector of all the selected columns that are not a \code{keys} element.
}
\description{
Get values names which are spread into columns.
}
\keyword{internal}
