% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{get_relabel_call}
\alias{get_relabel_call}
\title{Create relabel call from named character}
\usage{
get_relabel_call(labels)
}
\arguments{
\item{labels}{(named \code{character})
where name is name is function argument name and value is a function argument value.}
}
\value{
\code{call} object with relabel step.
}
\description{
Function creates relabel call from named character.
}
\examples{
get_relabel_call(
  labels = c(
    x = as.name("ANL"),
    AGE = "Age",
    AVAL = "Continuous variable"
  )
)

get_relabel_call(
  labels = c(
    AGE = "Age",
    AVAL = "Continuous variable"
  )
)
}
