% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{get_relabel_cols}
\alias{get_relabel_cols}
\title{Get columns to relabel}
\usage{
get_relabel_cols(columns_source, dplyr_call_data)
}
\arguments{
\item{columns_source}{(\code{list})}

\item{dplyr_call_data}{(\code{list})}
}
\value{
\code{columns_source} list without columns that have been reshaped.
}
\description{
Get columns to relabel excluding these which has been reshaped (pivot_wider).
}
\keyword{internal}
