% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_reshape_unite_col}
\alias{get_reshape_unite_col}
\title{Get unite columns}
\usage{
get_reshape_unite_col(selector)
}
\arguments{
\item{selector}{one element of selector_list obtained by \code{get_dplyr_call_data}.}
}
\value{
A \code{character} vector of all the selector's keys that are defined in the filters.
}
\description{
Get key names which spreads values into columns. Reshape is done only
on keys which are in \code{filter_spec}.
}
\keyword{internal}
