% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_reshape_unite_vals}
\alias{get_reshape_unite_vals}
\title{Get unite columns values}
\usage{
get_reshape_unite_vals(selector)
}
\arguments{
\item{selector}{one element of selector_list obtained by \code{get_dplyr_call_data}.}
}
\value{
A \code{character} vector of keys of the unite columns.
}
\description{
Get key values (levels) of the unite columns.
}
\keyword{internal}
