% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_module.R
\name{id_for_dataset}
\alias{id_for_dataset}
\title{Generate id for dataset}
\usage{
id_for_dataset(dataname)
}
\arguments{
\item{dataname}{(\code{character(1)}) the name of the dataset.}
}
\value{
\code{character(1)}.
}
\description{
Generate id for dataset
}
\keyword{internal}
