% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{is.delayed}
\alias{is.delayed}
\title{Is the specification resolved?}
\usage{
is.delayed(specification)
}
\arguments{
\item{specification}{Object to be evaluated.}
}
\value{
A single logical value.
}
\description{
Check that the specification is resolved against a given data source.
}
\examples{
is.delayed(1)
is.delayed(variables("df", "df"))
is.delayed(variables("df")) # Unknown selection
}
