% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_datanames.R
\name{is_single_dataset}
\alias{is_single_dataset}
\title{Verify uniform dataset source across data extract specification}
\usage{
is_single_dataset(...)
}
\arguments{
\item{...}{either \code{data_extract_spec} objects or lists of \code{data_extract_spec}
objects that do not contain \code{NULL}}
}
\value{
\code{TRUE} if all \code{data_extract_spec} objects come from the same dataset,
\code{FALSE} otherwise.
}
\description{
Checks if the input \code{data_extract_spec} objects all come from the same dataset.
}
