% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{list_extract_spec}
\alias{list_extract_spec}
\title{Make sure that the extract spec has list form}
\usage{
list_extract_spec(x, allow_null = FALSE)
}
\arguments{
\item{x}{\code{data_extract_spec} a single \code{data_extract_spec} or list of these}

\item{allow_null}{\code{logical} whether x can be \code{NULL}}
}
\value{
x as a list if it is not already
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
