% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-merge.R
\name{merge_expr}
\alias{merge_expr}
\title{Merge expression for selectors}
\usage{
merge_expr(
  selectors,
  output_name = "merged",
  join_fun = "dplyr::left_join",
  join_keys,
  allow_cartesian = FALSE
)
}
\arguments{
\item{selectors}{(\code{list} of \code{picks})}

\item{output_name}{(\code{character(1)})}

\item{join_fun}{(\code{character(1)}) name of the merge function.}

\item{join_keys}{(\code{join_keys})}
}
\description{
Merge expression for selectors
}
