% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{no_selected_as_NULL}
\alias{no_selected_as_NULL}
\title{Check select choices for no choice made}
\usage{
no_selected_as_NULL(x)
}
\arguments{
\item{x}{(\code{character}) Word that shall be checked for \code{NULL}, empty, "--no-selection".}
}
\value{
The word or \code{NULL}.
}
\description{
Check select choices for no choice made
}
