% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{parse_merge_key_i}
\alias{parse_merge_key_i}
\title{Parses merge keys}
\usage{
parse_merge_key_i(
  selector_list,
  idx,
  dplyr_call_data = get_dplyr_call_data(selector_list),
  merge_key = get_merge_key_i(selector_list, idx, dplyr_call_data)
)
}
\arguments{
\item{selector_list}{(\code{reactive})
output from \code{\link[=data_extract_multiple_srv]{data_extract_multiple_srv()}} or a reactive named list of
outputs from \code{\link[=data_extract_srv]{data_extract_srv()}}.
When using a reactive named list, the names must be identical to the shiny
ids of the respective
\code{\link[=data_extract_ui]{data_extract_ui()}}.}

\item{idx}{optional (\code{integer}) current selector index in all selectors list.}

\item{dplyr_call_data}{(\code{list}) simplified selectors with aggregated set of filters.}

\item{merge_key}{keys obtained from \code{get_merge_key_i}.}
}
\value{
\code{call} with merge keys.
}
\description{
Parses merge keys
}
\keyword{internal}
