% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-module_picks.R
\name{picks_module}
\alias{picks_module}
\alias{picks_ui}
\alias{picks_ui.list}
\alias{picks_ui.picks}
\alias{picks_srv}
\alias{picks_srv.list}
\alias{picks_srv.picks}
\title{Interactive picks}
\usage{
picks_ui(id, spec, container = "badge_dropdown")

\method{picks_ui}{list}(id, spec, container = "badge_dropdown")

\method{picks_ui}{picks}(id, spec, container = "badge_dropdown")

picks_srv(id = "", spec, data)

\method{picks_srv}{list}(id, spec, data)

\method{picks_srv}{picks}(id, spec, data)
}
\arguments{
\item{id}{(\code{character(1)}) Shiny module ID}

\item{spec}{(\code{picks} or \code{list}) Specification object created by \code{picks()} or a named list of such objects}

\item{container}{(\code{character(1)} or \code{function}) UI container type. Default is \code{"badge_dropdown"}.
Can also be one of \code{htmltools::tags} functions}

\item{data}{(\code{reactive}) Reactive expression returning the data object to be used for populating choices}
}
\value{
\itemize{
\item \code{picks_ui()}: UI elements for the input controls
\item \code{picks_srv()}: Server-side reactive logic returning the processed data
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates UI and server components for interactive \code{\link[=picks]{picks()}} in Shiny modules. The module is based on
configuration provided via \code{\link[=picks]{picks()}} and its responsibility is to determine relevant input
values

The module supports both single and combined specifications:
\itemize{
\item Single \code{picks} objects for a single input
\item Named lists of \code{picks} objects for multiple inputs
}
}
\details{
The module uses S3 method dispatch to handle different types of specifications:
\itemize{
\item \code{.picks} methods handle single specification
\item \code{.list} methods handle multiple specifications
}

The UI component (\code{picks_ui}) creates the visual elements, while the
server component (\code{picks_srv}) manages the reactive logic,
}
\seealso{
\code{\link[=picks]{picks()}} for creating specification objects
}
