% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-merge.R
\name{qenv_merge_selectors}
\alias{qenv_merge_selectors}
\title{Merge expression for selectors}
\usage{
qenv_merge_selectors(
  x,
  selectors,
  output_name = "merged",
  join_fun = "dplyr::left_join",
  allow_cartesian = TRUE
)
}
\arguments{
\item{x}{(\link[teal.data:teal_data]{teal.data::teal_data})}

\item{selectors}{(\verb{named list} of \code{picks})}

\item{output_name}{(\code{character(1)}) name of the merged dataset.}

\item{join_fun}{(\code{character(1)}) name of the merge function.}

\item{allow_cartesian}{(\code{logical(1)}) name of the merge function.}
}
\description{
Merge expression for selectors
}
