% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_delayed_datanames.R
\name{resolve_delayed_datanames}
\alias{resolve_delayed_datanames}
\title{Resolve \code{delayed_datanames}}
\usage{
resolve_delayed_datanames(des, datasets)
}
\arguments{
\item{des}{(\code{data_extract_spec} or \code{list} thereof) see \code{Details}}

\item{datasets}{(\code{character}) vector of dataset for which to resolve}
}
\value{
List of \code{delayed_data_extract_spec}s.
}
\description{
Convert \code{delayed_data_extract_spec}s containing \code{delayed_datanames} into normal ones.
}
\details{
This function is used internally.
}
\section{Resolution}{

\code{delayed_data_extract_specs} are resolved as follows:
\itemize{
\item \code{data_extract_specs} are returned as is
\item \code{delayed_data_extract_specs} where \code{dataname} is \code{character} are returned as is
\item \code{delayed_data_extract_specs} where \code{dataname} is \code{delayed_datanames} is first confronted
with names of datasets in the app and has its \code{datasets} attribute updated,
and then is converted to a list of \code{delayed_data_extract_spec}s of the same length as
the updated \code{datasets} attribute.
}
}

\keyword{internal}
