% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{resolver}
\alias{resolver}
\title{Resolve the specification}
\usage{
resolver(spec, data, ...)
}
\arguments{
\item{spec}{A object extraction specification.}

\item{data}{A \code{qenv()}, or \code{teal.data::teal_data()} object.}
}
\value{
A transform but resolved
}
\description{
Given the specification of some data to extract find if they are available or not.
}
\examples{
dataset1 <- datasets("df", function(x){head(x, 1)})
dataset2 <- datasets(is.matrix, function(x){head(x, 1)})
spec <- dataset1 & variables("a", "a")
td <- within(teal.data::teal_data(), {
  df <- data.frame(a = as.factor(LETTERS[1:5]), b = letters[1:5])
  m <- matrix()
})
resolver(dataset2, td)
resolver(spec, td)
spec <- dataset1 & variables("a", is.factor)
resolver(spec, td)
}
