% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.transform-package.R
\docType{package}
\name{teal.transform-package}
\alias{teal.transform}
\alias{teal.transform-package}
\title{\code{teal.transform}: Functions for extracting and merging data in the \code{teal} framework}
\description{
A standardized user interface for column selection, that facilitates dataset merging in 'teal' framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/teal.transform/}
  \item \url{https://github.com/insightsengineering/teal.transform/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.transform/issues}
}

}
\author{
\strong{Maintainer}: Dony Unardi \email{unardid@gene.com} [reviewer]

Authors:
\itemize{
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Nikolas Burkoff
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Junlue Zhao
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
