% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-tidyselect-helpers.R
\name{tidyselectors}
\alias{tidyselectors}
\alias{is_categorical}
\title{\code{tidyselect} helpers}
\usage{
is_categorical(min.len, max.len)
}
\arguments{
\item{min.len}{(\code{integer(1)}) minimal number of unique values}

\item{max.len}{(\code{integer(1)}) maximal number of unique values}
}
\description{
\code{tidyselect} helpers
}
\examples{
# select factor column but exclude foreign keys
variables(choices = is_categorical(min.len = 2, max.len = 10))

}
