% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-tm_merge.R
\name{tm_merge}
\alias{tm_merge}
\title{Merge module}
\usage{
tm_merge(label = "merge-module", picks, transformators = list())
}
\arguments{
\item{picks}{(\code{list} of \code{picks})
Example module}
}
\description{
Merge module
}
