% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Types}
\alias{Types}
\alias{datasets}
\alias{variables}
\alias{mae_colData}
\alias{values}
\title{Type specification}
\usage{
datasets(x, select = 1)

variables(x, select = 1)

mae_colData(x, select = 1)

values(x, select = 1)
}
\arguments{
\item{x}{Character specifying the names or functions to select them. The functions will be applied on the data or the names.}

\item{select}{Character of \code{x} or functions to select on x (only on names or positional not on the data of the variable).}
}
\value{
An object of the same class as the function with two elements: names the content of x, and select.
}
\description{
Define how to select and extract data
}
\section{Functions}{
\itemize{
\item \code{datasets()}: Specify datasets.

\item \code{variables()}: Specify variables.

\item \code{mae_colData()}: Specify colData.

\item \code{values()}: Specify values.

}}
\examples{
datasets("A")
c(datasets("A"), datasets("B"))
datasets(where(is.data.frame))
c(datasets("A"), variables(where(is.numeric)))
}
