% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{update_spec}
\alias{update_spec}
\title{Update a spec}
\usage{
update_spec(spec, type, value)
}
\arguments{
\item{spec}{A specification}

\item{type}{Which type was updated?}

\item{value}{What is the new selection?}
}
\value{
The specification with restored choices and selection if caused by the update.
}
\description{
Once a selection is made update the specification
}
