% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_datasets.R
\name{validate_keys_sufficient}
\alias{validate_keys_sufficient}
\title{Validates whether the provided keys are sufficient to merge the datasets slices}
\usage{
validate_keys_sufficient(join_keys, merged_selector_list)
}
\arguments{
\item{join_keys}{(\code{join_keys}) the provided join keys.}

\item{merged_selector_list}{(\code{list}) the specification of datasets' slices to merge.}
}
\value{
\code{TRUE} if the provided keys meet the requirement and \code{shiny}
validate error otherwise.
}
\description{
Validates whether the provided keys are sufficient to merge the datasets slices
}
\note{
The keys are not sufficient if the datasets slices described in
\code{merged_selector_list} come from datasets, which don't have the
appropriate join keys in \code{join_keys}.
}
\keyword{internal}
