% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_condition_range}
\alias{call_condition_range}
\title{\code{numeric} range condition call}
\usage{
call_condition_range(varname, range)
}
\arguments{
\item{varname}{(\code{name} or \code{character(1)})
name of the variable.}

\item{range}{(\code{numeric(2)})
range of the variable.}
}
\value{
\code{call}.
}
\description{
Compose \code{numeric} range condition call from inputs.
}
\keyword{internal}
