% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_extract_array}
\alias{call_extract_array}
\title{Get call to subset and select array}
\usage{
call_extract_array(dataname = ".", row = NULL, column = NULL, aisle = NULL)
}
\arguments{
\item{dataname}{(\code{character(1)} or \code{name}).}

\item{row}{(optional \code{name} or \code{call} or \code{logical} or \code{integer} or \code{character})
name of the \code{row} or condition.}

\item{column}{(optional \code{name} or \code{call} or \code{logical}, \code{integer} or \code{character})
name of the \code{column} or condition.}

\item{aisle}{(optional \code{name} or \code{call} or \code{logical} or \code{integer} or \code{character})
name of the \code{row} or condition.}
}
\value{
\code{\link[=Extract]{Extract()}} \code{call} for 3-dimensional array in \code{x[i, j, k]} notation.
}
\description{
Get call to subset and select array
}
\keyword{internal}
