% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_extract_list}
\alias{call_extract_list}
\title{Compose extract call with \code{$} operator}
\usage{
call_extract_list(dataname, varname, dollar = TRUE)
}
\arguments{
\item{dataname}{(\code{character(1)} or \code{name}) name of the object.}

\item{varname}{(\code{character(1)} or \code{name}) name of the slot in data.}

\item{dollar}{(\code{logical(1)}) whether returned call should use \code{$} or \code{[[} operator.}
}
\value{
\code{\link[=Extract]{Extract()}} \code{call} in \code{$} or \code{[[} notation (depending on parameters).
}
\description{
Compose extract call with \code{$} operator
}
\keyword{internal}
