% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_extract_matrix}
\alias{call_extract_matrix}
\title{Get call to subset and select matrix}
\usage{
call_extract_matrix(dataname = ".", row = NULL, column = NULL)
}
\arguments{
\item{dataname}{(\code{character(1)} or \code{name}).}

\item{row}{(\code{name} or \code{call} or \code{logical} or \code{integer} or \code{character}) optional
name of the \code{row} or condition.}

\item{column}{(\code{name} or \code{call} or \code{logical} or \code{integer} or \code{character}) optional
name of the \code{column} or condition.}
}
\value{
\code{\link[=Extract]{Extract()}} \code{call} for matrix in \code{x[i, j]} notation.
}
\description{
Get call to subset and select matrix
}
\keyword{internal}
