% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{calls_combine_by}
\alias{calls_combine_by}
\title{Combine calls by operator}
\usage{
calls_combine_by(operator, calls)
}
\arguments{
\item{operator}{(\code{character(1)} or \code{name})
name / symbol of the operator.}

\item{calls}{(\code{list} of calls)
list containing calls to be combined by \code{operator}}
}
\value{
a combined \code{call}
}
\description{
Combine list of calls by specific operator
}
\examples{
# use non-exported function from teal.transform
calls_combine_by <- getFromNamespace("calls_combine_by", "teal.transform")
call_condition_choice <- getFromNamespace("call_condition_choice", "teal.transform")
call_condition_range <- getFromNamespace("call_condition_range", "teal.transform")

calls_combine_by(
  "&",
  calls = list(
    call_condition_choice("SEX", "F"),
    call_condition_range("AGE", c(20, 50)),
    call_condition_choice("ARM", "ARM: A"),
    TRUE
  )
)
}
\keyword{internal}
