% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_no_multiple_selection}
\alias{check_no_multiple_selection}
\title{Checks that the \code{extract_input} specification does not allow multiple
selection}
\usage{
check_no_multiple_selection(extract_input)
}
\arguments{
\item{extract_input}{(\code{list} or \code{NULL}) a list of \code{data_extract_spec}}
}
\value{
Raises an error when check fails, otherwise, it returns \code{NULL}, invisibly.
}
\description{
Stops if condition not met.
}
