% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_single_module.R
\name{data_extract_single_ui}
\alias{data_extract_single_ui}
\title{Returns a \code{shiny.tag} with the UI elements for a \code{data_extract_spec}}
\usage{
data_extract_single_ui(id = NULL, single_data_extract_spec)
}
\arguments{
\item{id}{(\code{character(1)}) the id of the module.}

\item{single_data_extract_spec}{(\code{data_extract_spec}) the
\code{\link[=data_extract_spec]{data_extract_spec()}} object to handle.}
}
\value{
\code{shiny.tag} the HTML element defining the UI.
}
\description{
Returns a \code{shiny.tag} with the UI elements for a \code{data_extract_spec}
}
\details{
Creates a \code{shiny.tag} element defining the UI elements corresponding to a
single \code{data_extract_spec} object.
}
\keyword{internal}
