% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_datasets.R
\name{delayed_datasets}
\alias{delayed_datasets}
\title{Delayed datasets}
\usage{
delayed_datasets(x = "all")
}
\arguments{
\item{x}{(\code{character}) set of dataset names for wchich \code{delayed_data_extract_spec}s will be created;
set to \code{"all"} to use all available datasets}
}
\value{
Character string with \code{class} \code{delayed_datasets} and attribute \code{datasets}.
}
\description{
Generate \code{delayed_data_extract_spec} without prior knowledge of the data.
}
\details{
\code{delayed_datasets} is a character string with class \code{delayed_datasets}
and attribute \code{datasets}, which is set to \code{x}. The attribute specifies
a wishlist of datasets for which \code{delayed_data_extract_spec}s are to be created,
maintaining the same specification for \code{select}, \code{filter}, and \code{reshape}.

\code{delayed_data_extract_spec} that have \code{delayed_datasets} for \code{dataname} are resolved internally.

It is forbidden to use different \code{delayed_datasets} within one \code{delayed_data_extract_spec}
as well as to mix \code{delayed_datasets} with specific dataset specification within one \code{delayed_data_extract_spec}.
This is enforced when creating \code{data_extract_spec}s.
}
\section{Resolution}{

\code{delayed_data_extract_specs} are resolved as follows:
\itemize{
\item \code{data_extract_specs} are returned as is
\item \code{delayed_data_extract_specs} where \code{dataname} is \code{character} are returned as is
\item \code{delayed_data_extract_specs} where \code{dataname} is \code{delayed_datasets} is first confronted
with names of datasets in the app and has its \code{datasets} attribute updated,
and then is converted to a list of \code{delayed_data_extract_spec}s of the same length as
the updated \code{datasets} attribute.
}
}

\examples{
# resolve into delayed_data_extract_specs for all available datasets
data_extract_spec(
  dataname = delayed_datasets()
)

# resolve into delayed_data_extract_specs for available datasets from among ADSL and ADAE
data_extract_spec(
  dataname = delayed_datasets(c("ADSL", "ADAE"))
)

# use the same delayed_datasets() in child elements of a des
data_extract_spec(
  dataname = delayed_datasets(),
  select = select_spec(
    choices = variable_choices(
      data = delayed_datasets(),
      subset = function(data) names(Filter(is.numeric, data))
    ),
    selected = last_choice()
  )
)

}
