% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{determine}
\alias{determine}
\title{A method that should take a type and resolve it.}
\usage{
determine(type, data, ...)
}
\arguments{
\item{type}{The specification to resolve.}

\item{data}{The minimal data required.}
}
\value{
A list with two elements, the type resolved and the data extracted.
}
\description{
Generic that makes the minimal check on spec.
Responsible of subsetting/extract the data received and check that the type matches
}
\keyword{internal}
