% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-module_merge.R
\name{.merge_summary_list}
\alias{.merge_summary_list}
\title{Analyse selectors and concludes a merge parameters}
\usage{
.merge_summary_list(selectors, join_keys)
}
\value{
list containing:
\itemize{
\item mapping (\verb{named list}) containing selected values in each selector. This \code{mapping}
is sorted according to correct datasets merge order. \code{variables} contains names of the
variables in \code{ANL}
\item join_keys (\code{join_keys}) updated \code{join_keys} containing keys of \code{ANL}
}
}
\description{
Analyse selectors and concludes a merge parameters
}
\keyword{internal}
