% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merge_call.R
\name{get_anl_relabel_call}
\alias{get_anl_relabel_call}
\title{Gets the relabel call}
\usage{
get_anl_relabel_call(columns_source, datasets, anl_name = "ANL")
}
\arguments{
\item{columns_source}{(named \code{list})
where names are column names, values are labels + additional attribute \code{dataname}}

\item{datasets}{(named \code{list} of \code{reactive} or non-\code{reactive} \code{data.frame})
object containing data as a list of \code{data.frame}.
When passing a list of non-reactive \code{data.frame} objects, they are
converted to reactive \code{data.frame} objects internally.}

\item{anl_name}{(\code{character(1)})
Name of the analysis dataset.}
}
\value{
(\code{call}) to relabel \code{dataset} and assign to \code{anl_name}.
}
\description{
Gets the relabel call
}
