% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_datanames.R
\name{get_extract_datanames}
\alias{get_extract_datanames}
\title{Gets names of the datasets from a list of \code{data_extract_spec} objects}
\usage{
get_extract_datanames(data_extracts)
}
\arguments{
\item{data_extracts}{(\code{data_extract_spec(1)}) object or a list (of lists)
of \code{data_extract_spec}.}
}
\value{
\code{character} vector with the unique \code{dataname} set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Fetches \code{dataname} slot per \code{data_extract_spec} from a list of
\code{data_extract_spec}.
}
