% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_filter_call}
\alias{get_filter_call}
\title{Build a \code{dplyr} filter call}
\usage{
get_filter_call(filter, dataname = NULL, datasets = NULL)
}
\arguments{
\item{filter}{(\code{list}) Either list of lists or list with \code{select} and \code{selected} items.}

\item{dataname}{(\code{NULL} or \code{character}) name of dataset.}

\item{datasets}{(\code{NULL} or named \code{list}).}
}
\value{
\code{dplyr} filter \code{call}.
}
\description{
Build a \code{dplyr} filter call
}
\keyword{internal}
