% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_filter_module.R
\name{get_initial_filter_values}
\alias{get_initial_filter_values}
\title{Returns the initial values for the \code{vals} widget of a \code{filter_spec} object}
\usage{
get_initial_filter_values(filter, datasets)
}
\arguments{
\item{filter}{(\code{filter_spec}) the filter generated by a call to \code{\link[=filter_spec]{filter_spec()}}.}

\item{datasets}{(\verb{named list}) a list of reactive \code{data.frame} type objects.}
}
\value{
named \code{list} with two slots \code{choices} and \code{selected}.
}
\description{
Returns the initial values for the \code{vals} widget of a \code{filter_spec} object
}
\keyword{internal}
