% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dplyr_call.R
\name{get_select_call}
\alias{get_select_call}
\title{Parse \code{dplyr} select call}
\usage{
get_select_call(select)
}
\arguments{
\item{select}{(\code{character}) vector of selected column names.}
}
\value{
\code{dplyr} select \code{call}.
}
\description{
Parse \code{dplyr} select call
}
\keyword{internal}
