% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-delayed.R
\name{is.delayed}
\alias{is.delayed}
\title{Is the specification resolved?}
\usage{
is.delayed(x)
}
\arguments{
\item{x}{Object to be evaluated.}
}
\value{
A single logical value.
}
\description{
Check that the specification is resolved against a given data source.
}
\keyword{internal}
