% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_datasets.R
\name{merge_selectors}
\alias{merge_selectors}
\title{Merge selectors when \code{dataname}, \code{reshape}, \code{filters} and \code{keys} entries are identical}
\usage{
merge_selectors(selector_list)
}
\arguments{
\item{selector_list}{(\code{reactive})
output from \code{\link[=data_extract_multiple_srv]{data_extract_multiple_srv()}} or a reactive named list of
outputs from \code{\link[=data_extract_srv]{data_extract_srv()}}.
When using a reactive named list, the names must be identical to the shiny
ids of the respective
\code{\link[=data_extract_ui]{data_extract_ui()}}.}
}
\value{
List of merged selectors or original parameter if the conditions to merge are
not applicable.
}
\description{
Merge selectors when \code{dataname}, \code{reshape}, \code{filters} and \code{keys} entries are identical
}
\keyword{internal}
