% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{resolver}
\alias{resolver}
\title{Resolve the specification}
\usage{
resolver(spec, data)
}
\arguments{
\item{spec}{A object extraction specification.}

\item{data}{The qenv where the specification is evaluated.}
}
\value{
A transform but resolved
}
\description{
Given the specification of some data to extract find if they are available or not.
The specification for selecting a variable shouldn't depend on the data of said variable.
}
\examples{
dataset1 <- datasets(is.data.frame)
dataset2 <- datasets(is.matrix)
spec <- dataset1 & variables("a", "a")
td <- within(teal.data::teal_data(), {
  df <- data.frame(a = as.factor(LETTERS[1:5]), b = letters[1:5])
  df2 <- data.frame(a = LETTERS[1:5], b = 1:5)
  m <- matrix()
})
resolver(spec | dataset2, td)
resolver(dataset2, td)
resolver(spec, td)
spec <- dataset1 & variables("a", is.factor)
resolver(spec, td)
}
