% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-resolver.R
\name{resolver}
\alias{resolver}
\title{Resolve \code{picks}}
\usage{
resolver(x, data)
}
\arguments{
\item{x}{(\code{\link[=picks]{picks()}}) settings for picks.}

\item{data}{(\code{\link[teal.data:teal_data]{teal.data::teal_data()}} \code{environment} or \code{list}) any data collection supporting object extraction with \code{[[}.
Used to determine values of unresolved \code{picks}.}
}
\value{
resolved \code{picks}.
}
\description{
Resolve iterates through each \code{picks} element and determines values .
}
\examples{
x <- picks(datasets(tidyselect::where(is.data.frame)), variables("a", "a"))
data <- list(
  df1 = data.frame(a = as.factor(LETTERS[1:5]), b = letters[1:5]),
  df2 = data.frame(a = LETTERS[1:5], b = 1:5),
  m = matrix()
)
resolver(x = x, data = data)
}
