% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-resolver.R
\name{resolver}
\alias{resolver}
\title{Resolve the specification}
\usage{
resolver(x, data)
}
\arguments{
\item{data}{The qenv where the specification is evaluated.}

\item{spec}{A object extraction specification.}
}
\value{
A specification but resolved: the names and selection is the name of the objects (if possible).
}
\description{
Given the specification of some data to extract find if they are available or not.
The specification for selecting a variable shouldn't depend on the data of said variable.
}
\examples{
dataset1 <- datasets(where(is.data.frame))
dataset2 <- datasets(where(is.matrix))
spec <- c(dataset1, variables("a", "a"))
td <- within(teal.data::teal_data(), {
  df <- data.frame(a = as.factor(LETTERS[1:5]), b = letters[1:5])
  df2 <- data.frame(a = LETTERS[1:5], b = 1:5)
  m <- matrix()
})
resolver(list(spec, dataset2), td)
resolver(dataset2, td)
resolver(spec, td)
spec <- c(dataset1, variables("a", where(is.character)))
resolver(spec, td)
}
