% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-resolver.R
\name{resolver}
\alias{resolver}
\title{Resolve \code{picks}}
\usage{
resolver(x, data)
}
\arguments{
\item{data}{(\code{\link[teal.data:teal_data]{teal.data::teal_data()}} \code{environment} or \code{list}) any data collection supporting object extraction with \code{[[}.
Used to determine values of unresolved \code{picks}.}

\item{picks}{(\code{\link[=picks]{picks()}}) settings for picks.}
}
\value{
resolved \code{picks}.
}
\description{
Resolve iterates through each \code{picks} element and determines values .
}
\examples{
# todo: fix example to use environment or a list
x1 <- datasets(where(is.data.frame))
x2 <- picks(x1, variables("a", "a"))
data <- within(teal.data::teal_data(), {
  df <- data.frame(a = as.factor(LETTERS[1:5]), b = letters[1:5])
  df2 <- data.frame(a = LETTERS[1:5], b = 1:5)
  m <- matrix()
})
resolver(x = x1, data = data)
resolver(x = x2, data = data)
}
