% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.transform-package.R
\docType{package}
\name{teal.transform-package}
\alias{teal.transform}
\alias{teal.transform-package}
\title{teal.transform: Functions for extracting and merging data in the \code{teal} framework}
\description{
Functions for extracting and merging data in the teal framework.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
