% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-tidyselect-helpers.R
\name{tidyselectors}
\alias{tidyselectors}
\alias{is_key}
\alias{is_categorical}
\alias{no_more_choices_than}
\title{\code{tidyselect} helpers}
\usage{
is_key()

is_categorical(max.len, min.len)

no_more_choices_than(max.len)
}
\arguments{
\item{max.len}{(\code{integer(1)}) maximal number of unique values}

\item{min.len}{(\code{integer(1)}) minimal number of unique values}
}
\description{
\code{tidyselect} helpers
}
\examples{
# select keys (primary and foreign)
variables(choices = is_key())

# select factor column but exclude foreign keys
variables(choices = where(~ is.factor(.x) & !is_foreign_key()))
}
