% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Types}
\alias{Types}
\alias{datasets}
\alias{variables}
\alias{mae_colData}
\alias{mae_sampleMap}
\alias{mae_experiments}
\alias{values}
\title{Type specification}
\usage{
datasets(x, select = first)

variables(x, select = first)

mae_colData(x, select = first)

mae_sampleMap(x, select = first)

mae_experiments(x, select = first)

values(x, select = first)
}
\arguments{
\item{x}{Character specifying the names or functions to select them. The functions will be applied on the data or the names.}

\item{select}{Character of \code{x} or functions to select on x (only on names or positional not on the data of the variable).}
}
\value{
An object of the same class as the function with two elements: names the content of x, and select.
}
\description{
Define how to select and extract data
}
\section{Functions}{
\itemize{
\item \code{datasets()}: Specify datasets.

\item \code{variables()}: Specify variables.

\item \code{mae_colData()}: Specify variables of MultiAssayExperiment col Data.

\item \code{mae_sampleMap()}: Specify variables of MultiAssayExperiment sampleMap.

\item \code{mae_experiments()}: Specify variables of MultiAssayExperiment experiments.

\item \code{values()}: Specify values.

}}
\examples{
datasets("A")
datasets("A") | datasets("B")
datasets(is.data.frame)
datasets("A") & variables(is.numeric)
}
