% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-picks.R
\name{Types}
\alias{Types}
\alias{picks}
\alias{values}
\alias{mae_colData}
\title{Choices/selected settings}
\usage{
picks(...)

values(
  choices = tidyselect::everything(),
  selected = tidyselect::everything(),
  multiple = TRUE,
  fixed = !.is_tidyselect(choices) && length(choices) == 1,
  ...
)

mae_colData(choices = tidyselect::everything(), selected = 1, multiple = FALSE)
}
\arguments{
\item{...}{additional arguments delivered to \code{pickerInput}}

\item{choices}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}} or \code{character}>
One unquoted expression to be used to picks the choices.}

\item{selected}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}} or \code{character}>
One unquoted expression to be used to picks from choices to be selected.}

\item{multiple}{<\code{logical(1)}> if more than one selection is possible.}

\item{fixed}{<\code{logical(1)}> selection will be fixed and not possible to change interactively.}

\item{keep_order}{<\code{logical(1)}> if the selected should follow the selection order. If \code{FALSE}
\code{selected} returned from \code{srv_module_input()} would be ordered according to order in \code{choices}.}
}
\value{
\code{picks} object containing specified settings
}
\description{
Define choices and default selection of variables from datasets.
}
\section{Functions}{
\itemize{
\item \code{picks()}: specify a selector.

\item \code{values()}: Specify variables.

\item \code{mae_colData()}: Specify colData.

}}
\examples{

# Initialize selector for `iris` to select columns between `Sepal.Length` and `Petal.Width`
#  with first
picks(
  datasets(choices = "iris"),
  variables(choices = Sepal.Length:Petal.Width, selected = 1)
)
picks(
  datasets(choices = c("iris", "mtcars"), selected = "iris"),
  variables(choices = tidyselect::everything(), selected = 1)
)
picks(
  datasets(choices = c("iris", "mtcars"), selected = 1),
  variables(choices = tidyselect::where(is.numeric), selected = 1)
)
picks(
  datasets(choices = tidyselect::everything(), selected = 1),
  variables(choices = is_categorical(min.len = 2, max.len = 15), selected = 1:2)
)
}
