% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-types.R
\name{Types}
\alias{Types}
\alias{datasets}
\alias{variables}
\alias{mae_colData}
\alias{values}
\title{Type specification}
\usage{
datasets(choices = tidyselect::everything(), selected = 1)

variables(choices = tidyselect::everything(), selected = 1)

mae_colData(choices = tidyselect::everything(), selected = 1)

values(choices = tidyselect::everything(), selected = 1)
}
\arguments{
\item{choices}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One unquoted expression to be used to pick the choices.}

\item{selected}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One unquoted expression to be used to pick from choices to be selected.}
}
\value{
An object of the same class as the function with two elements: names the content of x, and select.
}
\description{
Define how to select and extract data
}
\section{Functions}{
\itemize{
\item \code{datasets()}: Specify datasets.

\item \code{variables()}: Specify variables.

\item \code{mae_colData()}: Specify colData.

\item \code{values()}: Specify values.

}}
\examples{
datasets()
datasets("A")
c(datasets("A"), datasets("B"))
datasets(where(is.data.frame))
c(datasets("A"), variables(where(is.numeric)))
}
